csrf token validation before deleting the image

+++ src/Controller/ProjectController.php
@@ -180,15 +180,19 @@ class ProjectController extends AbstractController
 
     public function deleteImage(Image $image, Request $request)
     {
-        unlink($this->getParameter('images_dir') . $image->getUniqueName());
+        $csrf_token = $request->request->get('csrf_token');
+        if ($this->isCsrfTokenValid('image_deletion_' . $image->getId(), $csrf_token))
+        {
+            unlink($this->getParameter('images_dir') . $image->getUniqueName());
 
-        $project = $image->getProject();
-        $project->removeImage($image);
-        $this->em->remove($image);
-        $this->em->flush();
+            $project = $image->getProject();
+            $project->removeImage($image);
+            $this->em->remove($image);
+            $this->em->flush();
 
-        // Flash message
-        $this->addFlash('info', 'Image supprimée avec succés !');
+            // Flash message
+            $this->addFlash('info', 'Image supprimée avec succés !');
+        }
 
         // Redirection
         return $this->redirectToRoute('app_project_update', [
